more generic kfproducer
